requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
python-docx==1.2.0
pdfplumber==0.11.7
dateparser==1.2.2
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
